REMOVED NowPlaying package & REMOVED PopularMovie package.. renamed Media Components to Movies Component
