+++ src/component/Media/MovieView.js
@@ -1,25 +1,24 @@
 import React from 'react';
-import {IMG_URL} from "../Constants/constants";
-import {Lightbox} from "primereact/components/lightbox/Lightbox";
 import {ScrollPanel} from "primereact/components/scrollpanel/ScrollPanel";
-import {Fieldset} from "primereact/components/fieldset/Fieldset";
 import {Card} from "primereact/components/card/Card";
+import {Fieldset} from "primereact/components/fieldset/Fieldset";
 import SimilarMediaComponent from "../SimilarMedia/SimilarMediaComponent";
+import {IMG_URL} from "../Constants/constants";
 
-const NowPlayingMovieView = props => {
+const MovieView = props => {
 
     const renderMovieReviews = () => {
-        let reviews = props.nowPlayingMovieReviews;
+        let reviews = props.movieReviews;
         if(reviews.length === 0){
             return(
-                <div className={'p-grid movie-noReviews'}>
-                    <Card subTitle={"Reviews"} className={'movie-reviewsHeader movie-reviewText p-col-12 nowPlayingMovie-text'}>
+                <div className={'p-grid movie-noReviews movie-text'}>
+                    <Card subTitle={"Reviews"} className={'movie-reviewsHeader movie-reviewText p-col-12'}>
                         There are no reviews at this time
                     </Card>
                 </div>
             )
         }
-        var allReviews = reviews.map((review, key) => {
+        let allReviews = reviews.map((review, key) => {
             return(
                 <div key={key} className="content-section introduction">
                     <div className="feature-intro">
@@ -28,7 +27,7 @@ const NowPlayingMovieView = props => {
                                   toggleable={true}
                                   collapsed={true}>
                             <ScrollPanel className={"movie-reviewScrollPanel"}>
-                                <div className={'movie-reviewContent nowPlayingMovie-text'}>
+                                <div className={'movie-reviewContent movie-text'}>
                                     {review.content}
                                 </div>
                             </ScrollPanel>
@@ -55,10 +54,10 @@ const NowPlayingMovieView = props => {
 
     const renderMovieOverview = (event) => {
         return(
-            <div className="content-section implementation scrollpanel-demo nowPlayingMovie-text">
+            <div className="content-section implementation scrollpanel-demo movie-text">
                 <Card subTitle={"Overview"}>
                     <ScrollPanel className={"movie-overview"}>
-                        {props.nowPlayingMovie.overview}
+                        {props.movie.overview}
                     </ScrollPanel>
                 </Card>
             </div>
@@ -66,39 +65,39 @@ const NowPlayingMovieView = props => {
     }
 
     return(
-        <div className={'p-grid p-nogutter p-col-12 p-align-stretch vertical-container nowPlayingMovie'}>
+        <div className={'p-grid p-nogutter p-col-12 p-align-stretch vertical-container movie'}>
             <div className={'p-grid p-nogutter p-col-12 '}>
                 <div className={'p-col-5'}>
                     <div className={'movie-imageContainer'}>
-                        <Lightbox type={'content'}>
+                        {/*<Lightbox type={'content'}>*/}
                             <a className={'group'} >
-                                <img src={`${IMG_URL}${props.nowPlayingMovie.poster_path}`} alt={props.nowPlayingMovie.original_title} className={'image-movie-poster'}/>
+                                <img src={`${IMG_URL}${props.movie.poster_path}`} alt={props.movie.original_title} className={'image-movie-poster'}/>
                             </a>
-                            <div>
-                                <iframe title="Video"
-                                        width="560"
-                                        height="315"
-                                        src={"https://www.youtube.com/embed/" + props.trailer}
-                                        frameBorder="0"
-                                        allowFullScreen>
+                            {/*<div>*/}
+                                {/*<iframe title="Video"*/}
+                                        {/*width="560"*/}
+                                        {/*height="315"*/}
+                                        {/*src={"https://www.youtube.com/embed/" + props.trailer}*/}
+                                        {/*frameBorder="0"*/}
+                                        {/*allowFullScreen>*/}
 
-                                </iframe>
-                            </div>
-                        </Lightbox>
+                                {/*</iframe>*/}
+                            {/*</div>*/}
+                        {/*</Lightbox>*/}
                         <div className={'p-col-12'}>
                             {renderMovieOverview()}
                         </div>
                     </div>
                 </div>
                 <div className={'p-col-7'}>
-                    <div className={'movie-title nowPlayingMovie-text'}>
-                        {props.nowPlayingMovie.title}
+                    <div className={'movie-title movie-text'}>
+                        {props.movie.title}
                     </div>
                     <div className={'p-grid p-nogutter p-col-12'} style={{textAlign: 'center'}}>
                         <div className={"movie-commentsAndReviews"}>
-                            {props.nowPlayingMovieReviews.length !== 0 ?
-                                <Card subTitle="Reviews" className={'movie-reviewsHeader nowPlayingMovie-text'}>
-                                    <div className={'movie-reviewClickText nowPlayingMovie-text'}>
+                            {props.movieReviews.length !== 0 ?
+                                <Card subTitle="Reviews" className={'movie-reviewsHeader movie-text'}>
+                                    <div className={'movie-reviewClickText'}>
                                         Click the name to show review
                                     </div>
                                 </Card>
@@ -109,12 +108,12 @@ const NowPlayingMovieView = props => {
                     </div>
                 </div>
             </div>
-            {props.similarMedia.length != 0 ?
+            { props.similarMedia.length !== 0 ?
                 <div className={'similar-margin-top-10'}>
                     <div className={'p-grid p-col-12 p-nogutter similar-margin-bottom-5'}>
                         <div className={'p-offset-1 p-col-10'}>
-                            <SimilarMediaComponent id={props.nowPlayingMovie.id}
-                                                   title={props.nowPlayingMovie.title}
+                            <SimilarMediaComponent id={props.movie.id}
+                                                   title={props.movie.title}
                                                    refreshMoviePage={props.refreshMoviePage}
                                                    similarMedia={props.similarMedia}
                             />
@@ -127,4 +126,4 @@ const NowPlayingMovieView = props => {
     )
 
 }
-export default NowPlayingMovieView;
+export default MovieView;
