--- src/component/NowPlaying/NowPlayingPanel.js
@@ -1,18 +1,0 @@
-import React, {Component} from 'react'
-import NowPlayingComponent from "./NowPlayingComponent";
-
-class NowPlayingPanel extends Component{
-    constructor(props){
-        super(props);
-
-    }
-    render(){
-        return(
-            <div>
-                {/*<NowPlaying />*/}
-                <NowPlayingComponent />
-            </div>
-        )
-    }
-}
-export default NowPlayingPanel;
