--- src/component/NowPlaying/css/nowPlaying.css
@@ -1,68 +1,0 @@
-.nowPlaying{
-    background: black;
-}
-
-.image-poster{
-    width: 100%;
-    height: 60vh;
-}
-
-.list-text-container{
-    margin: 2%;
-    text-align: left;
-    font-size: 20px;
-    max-width: 90%;
-}
-
-.list-text{
-    margin: 2%;
-}
-
-.list-text b{
-    font-size: 17px;
-}
-
-.overview-container{
-    margin-left: 5%;
-}
-
-.list-overview-scrollPane{
-    font-size: 15px;
-    text-align: left;
-    width: 100%;
-    height: 25vh;
-}
-
-.list-text-overview {
-    font-size: 17px;
-}
-
-.list-text-overview b{
-    margin-left: 5%;
-}
-
-.dataview-listItem {
-    padding: 2em;
-    border-bottom: 1px solid #222;
-    width: 10px;
-}
-
-.grid-header{
-    font-size: 18px;
-    max-height: 5px;
-    margin-bottom: 30px;
-    margin-left: 10%;
-}
-
-.nowPlaying-text{
-    color: white;
-}
-
-a {
-    transition: all 0.7s;
-}
-
-a:hover{
-    text-decoration: none;
-    color: white;
-}
