--- src/component/NowPlaying/css/nowPlayingMovie.css
@@ -1,7 +1,0 @@
-.nowPlayingMovie{
-    background: black;
-}
-
-.nowPlayingMovie-text{
-    color: white;
-}
